+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Fix unwanted timer submitting (!906)
 - Fix minimum allowed value in the poll timer. (!897)
 - Fix unresponsive hang up button on participant leave. (!903)
+- Fix cross icon showing on timer without ready check (!919)
 
 ### Stability Improvements